+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -15,6 +15,7 @@ import dev.yanianz.essentials.customscreens.CustomScreenModule;
 import dev.yanianz.essentials.polls.PollsModule;
 import dev.yanianz.essentials.nicknames.NicknamesModule;
 import dev.yanianz.essentials.bubbles.ChatBubblesModule;
+import dev.yanianz.essentials.network.BungeeChatModule;
 import dev.yanianz.essentials.chatcustomization.ChatCustomizationModule;
 import dev.yanianz.essentials.reports.ReportsModule;
 import dev.yanianz.essentials.notes.NotesModule;
@@ -105,6 +106,7 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(ReportsModule.class, new ReportsModule(this.plugin));
         this.modules.put(ChatCustomizationModule.class, new ChatCustomizationModule(this.plugin));
         this.modules.put(ChatBubblesModule.class, new ChatBubblesModule(this.plugin));
+        this.modules.put(BungeeChatModule.class, new BungeeChatModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }